requirements.txt
crawl4ai==1.0.0
flask==2.0.1
requests==2.25.1
pyyaml==6.0
youtube_transcript_api==0.5.0
beautifulsoup4==4.10.0
lxml==4.9.1
PyQt5==5.15.6
PyQtWebEngine==5.15.5
pandas==1.3.5
matplotlib==3.5.1
nltk==3.6.7
pillow==9.0.1
